add client list feature; rename all screen files
