+++ lib/utils/utils.dart
@@ -1,6 +1,13 @@
+import 'package:flutter/material.dart';
 import 'package:mask_text_input_formatter/mask_text_input_formatter.dart';
 
 const String CLIENT_BASE_KEY = "CLIENT_BASE_KEY";
 
 var phoneFormatter = new MaskTextInputFormatter(
     mask: '(##) #####-####', filter: {"#": RegExp(r'[0-9]')});
+
+var buttonStyle = ButtonStyle(
+    shape: MaterialStateProperty.all<RoundedRectangleBorder>(
+        RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(8),
+            side: BorderSide(color: Colors.black))));
